requirements.txt
Django
django-filter
djangorestframework
django-cors-headers
django-debug-toolbar
django-mptt
psycopg2
pytest
pytest-django
google-api-python-client
requests